Added Important Stolen Code Comments to Roman Numeral Related Code
A comment was added to romanNumeralClock.py and romeNumConv.py,
indicating a chunk of code used in both of these was not made by the
author of the code files. This comment recognizes the fact that
someone else made the code which was then taken by KesseNones from
StackOverflow some two years ago. This comment will eventually be
deleted when an original code block exists there.
Also, a small comment was added to both files describing what the
rome_conv function does.