Created GUI Version of Grade Calculator
The file gradeCalcGUI.py was created which is a GUI version of
what the original grade calculator does. This version is easier to
use and more convenient for calculating a bunch of grades back to back.
The file cPrograms.py was also modified to be better and to include the
new GUI weighted grade calculation program.