Added Comments to binEncoder.py and Changed its Font
Comments were added to the file binEncoder.py to increase readability.
The font was also decreased for the input fields and the output was
made an entry. The decreased font size was to make the program more
friendly for smaller displays. The changing of output from a label
to entry was to allow for copying to the clipboard to occur. Lastly,
the automatic copy to clipboard was removed for the 
decrease of program intrusion to user function. A homemade library was
also imported to make the base conversions work.