Added Comments to binaryClock.py
Comments were added to the binary clock program file to make it more
readable. A line for string concatenation was replaced with a formatted
string for an easier understanding of the string composition. The font
of the output time was also decreased to make the program friendlier
for smaller displays.