Added Comments to and Improved baseCounter.py
Comments were added to the base stopwatch program to increase
understanding of the code. The speed factor slider was changed to a
precision slider. The precision slider controls the time precision
of the second counter. An edge case was also corrected that caused
the time total to keep increasing when the pause button was repeatedly
pressed. The time output also now exists in an entry as apposed to a
label, making the output more stable and copy-pasteable.