Improved Floating Point Abilities of baseCalculator.py
The code was heavily reworked to support floating point arithmetic
for bases 2 through 36 in the four basic operations of addition,
subtraction, multiplication, and division. The floating points behave
strange in places since a lot of base conversions are involved. There
are also edge cases involving floats that are too small or too big.
Generally, floating points now work in arithmetic of different bases
mostly. There are future plans to try and make it work even better.