Added Comments to coinFlipGUI.py and Improved It
Comments were added to the coin flipping program
to increase readability. The program itself was also refactored to make
it easier to tell separate flips apart from each other with the use of
a progress statement and an artificial wait. Essentially,
the user clicks the flip coin button, "Flipping..." is displayed for 1
second, then the result is displayed. The output of the flip was also
turned into an entry to make it copy-able and also to make the output
display more consistently. Beyond these changes,
some coding optimizations were made
to make the code look less like garbage.