Greatly Refactored Easter Calculator
The code of the Easter Calculator was remade to be far less gross
looking and to also work better, taking advantage of stuff already
made else where. It's still not 100 percent accurate but much
better than it previously was. It took some adjustment for the moon
phase base for the metric time to make it accurate again.
Now this calculator works without the code looking like complete
garbage.