Refactored Reformed Calendar to be way less crummy
The code was changed to only display the reformed calendar date
string and nothing else that was cluttering output.
The code was also greatly changed to be way shorter and functional
by taking advantage of external libraries and slick list indexing
tricks. What resulted was a program of about 100 lines that does
the same thing but probably more efficiently. Perhaps a future
program may be made that had the cycle information that was removed
from this version of the calendar.