Refactored seasonal calendar calculator greatly
This calendar does the same thing but was greatly optimized behind
the scenes to be less ass. Also, the font size was decreased and
the UI was shifted to be smaller and less imposing. Overall though
this was mostly performing optimizations to make the code way less
terrible than it was. It also has comments now too.