Created New Version of Reformed Calendar Calculator
This version of the reformed calendar calculator displays the
resulting calendar page of the input date. This program also
displays the current reformed calendar date page when booted up.
When the user enters a valid date, the calendar date in the reformed
calendar is calculated and displayed in the same way the live one
is displayed.