Made Error Handling of Input More Thorough
The input fetch now checks to see if the input number is a valid
float. If it is, the string given is returned, otherwise the string
"0" is given instead because it was invalid. This makes some of the
background parsing less annoying. This may also be applied for some
other input parsing code becuase it's more elegant.