Added Executable Version of Omniprogram
It's now possible to run the Omniprogram without needing Python
installed. To run this, download the omniprogramExecutableVersion
folder, unzip it, go into the dist subfolder, and run the exe.
This was built on Windows so it probably won't work on Unix based
systems like Mac or Linux Distros. I may try using a VM to build
those comptatable versions later.