Fixed Super Broken Asimov Metric Time Calculator
The calculator's been mostly fixed now! There is still an issue
with timezones that I'm not bothering to touch right now but
otherwise it's *much* better than it was. You just have to assume
the time you're inputting will get shifted to UTC so that can be
tricky. I'll figure out the timezone stuff later.